refactor(queries): move auth, account and billing hooks into the queries layer
