+++ src/components/Auth/SignUp.tsx
@@ -15,7 +15,7 @@ import { Controller, FormProvider, useForm } from 'react-hook-form'
 import { Trans, useTranslation } from 'react-i18next'
 import { Navigate, NavLink, useOutletContext } from 'react-router-dom'
 import { useAnalytics } from '~components/AnalyticsProvider'
-import { IRegisterParams } from '~components/Auth/authQueries'
+import { IRegisterParams } from '~queries/auth'
 import { useAuth } from '~components/Auth/useAuth'
 import { VerificationPending } from '~components/Auth/Verify'
 import { default as InputBasic } from '~components/Form/InputBasic'
