+++ src/components/Integrator/NotIntegratorNotice.tsx
@@ -5,7 +5,7 @@ import { LuSquareStack } from 'react-icons/lu'
 import { getApiErrorMessage } from '~components/Auth/api'
 import { useAuth } from '~components/Auth/useAuth'
 import { LocalStorageKeys } from '~components/Auth/useAuthProvider'
-import { useOrganizationNames } from '~components/Dashboard/Menu/useOrganizationNames'
+import { useOrganizationNames } from '~queries/organization'
 import { useSelectOrganization } from '~components/Dashboard/Menu/useSelectOrganization'
 import DeleteModal from '~components/Modal/DeleteModal'
 import { useToast } from '~components/Toast'
