+++ src/components/Pricing/Plans.tsx
@@ -1,5 +1,4 @@
 import { Flex, Progress, SimpleGrid, Tabs, Tag } from '@chakra-ui/react'
-import { useQuery } from '@tanstack/react-query'
 import { useCallback, useMemo } from 'react'
 import { FormProvider, useForm } from 'react-hook-form'
 import { Trans, useTranslation } from 'react-i18next'
@@ -13,71 +12,16 @@ import {
   LuUsers,
   LuVote,
 } from 'react-icons/lu'
-import { ApiEndpoints } from '~components/Auth/api'
 import { useSubscription } from '~components/Auth/Subscription'
-import { useAuth } from '~components/Auth/useAuth'
 import { ListStateAlert } from '~components/Feedback/ListStateAlert'
 import { getPlanKey, isPlanNamed, PlanName } from '~constants'
-import { QueryKeys } from '~src/queries/keys'
+import { usePlans, type Plan, type SubscriptionCheckoutFormValues } from '~queries/plans'
 import PricingCard from './Card'
 import { useSubscriptionCheckout } from './use-subscription-checkout'
 
-export type Plan = {
-  id: string
-  name: string
-  stripeID: string
-  yearlyPrice: number
-  monthlyPrice: number
-  default: boolean
-  freeTrialDays: number
-  organization: {
-    teamMembers: number
-    subOrgs: number
-    censusSize: number
-    maxProcesses: number
-    maxCensus: number
-    maxDuration: string
-    customURL: boolean
-    drafts: boolean
-  }
-  votingTypes: {
-    single: boolean
-    multiple: boolean
-    approval: boolean
-    cumulative: boolean
-    ranked: boolean
-    weighted: boolean
-  }
-  features: {
-    anonymous: boolean
-    overwrite: boolean
-    liveResults: boolean
-    personalization: boolean
-    emailReminder: boolean
-    '2FAsms': number
-    '2FAemail': number
-    whiteLabel: boolean
-    liveStreaming: boolean
-    phoneSupport: boolean
-  }
-}
-
-export type SubscriptionCheckoutFormValues = {
-  billingPeriod: 'month' | 'year'
-  planId: string | null
-}
-
-export const usePlans = () => {
-  const { bearedFetch } = useAuth()
-  return useQuery({
-    queryKey: QueryKeys.plans,
-    queryFn: () => bearedFetch<Plan[]>(ApiEndpoints.Plans),
-    // Sort by price
-    select: (data) => data.sort((a, b) => a.yearlyPrice - b.yearlyPrice),
-    // Cache for 20 minutes
-    staleTime: 20 * 60 * 1000,
-  })
-}
+// Re-exported for existing consumers; the canonical definitions live in the queries layer.
+export { usePlans }
+export type { Plan, SubscriptionCheckoutFormValues }
 
 export const usePlanTranslations = (plans?: Plan[]) => {
   const { t } = useTranslation()
