+++ src/queries/storage.ts
@@ -1,0 +1,18 @@
+import { useMutation } from '@tanstack/react-query'
+import { ApiEndpoints } from '~components/Auth/api'
+import { useAuth } from '~components/Auth/useAuth'
+
+export const useUploadFile = () => {
+  const { bearedFetch } = useAuth()
+  return useMutation({
+    mutationFn: async (file: File) => {
+      const formData = new FormData()
+      formData.append('file1', file)
+      const response = await bearedFetch<{ urls: string[] }>(ApiEndpoints.Storage, {
+        method: 'POST',
+        body: formData,
+      })
+      return response.urls[0]
+    },
+  })
+}
